fix: extract duplicate methods
